pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ga4gh-sidecar"
version = "0.1.0"
description = "A reverse proxy sidecar for GA4GH service-info interception and dynamic metadata merging"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Karanjot Singh", email = "karanjots801@gmail.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "deepmerge>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
ga4gh-sidecar = "ga4gh_sidecar.main:cli"

[project.entry-points."ga4gh_sidecar.plugins"]
tes = "ga4gh_sidecar.plugins.tes:TESPlugin"
wes = "ga4gh_sidecar.plugins.wes:WESPlugin"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
